+++ src/Support/Facades/PostalCodes.php
@@ -2,6 +2,7 @@
 
 namespace Axlon\PostalCodeValidation\Support\Facades;
 
+use Axlon\PostalCodeValidation\Support\Constraint;
 use Illuminate\Support\Facades\Facade;
 
 /**
@@ -14,10 +15,32 @@ use Illuminate\Support\Facades\Facade;
 class PostalCodes extends Facade
 {
     /**
+     * Get a postal_code constraint builder instance.
+     *
+     * @param string|string[] ...$parameters
+     * @return \Axlon\PostalCodeValidation\Support\Constraint
+     */
+    public static function for(...$parameters): Constraint
+    {
+        return new Constraint('postal_code', $parameters);
+    }
+
+    /**
      * @inheritDoc
      */
     protected static function getFacadeAccessor(): string
     {
         return 'postal_codes';
+    }
+
+    /**
+     * Get a postal_code_with constraint builder instance.
+     *
+     * @param string|string[] ...$parameters
+     * @return \Axlon\PostalCodeValidation\Support\Constraint
+     */
+    public static function with(...$parameters): Constraint
+    {
+        return new Constraint('postal_code_with', $parameters);
     }
 }